Set up fractal to work with npm / dotnet commands
Handful of changes to get this to work.

Firstly, in order to get fractal to play nice on Windows, the
`fractal.js` file needed to be renamed and a reference added to the
`package.json` to tell fractal where to find the config file.
https://github.com/frctl/fractal/issues/118#issuecomment-255254117

Added quite a few changes to the npm scripts.

- Changed `npm run bundle:dev` to only compile static assets in
development mode, whereas `npm run bundle:watch` will compile static
assets in development mode and then watch for any changes to trigger a
new build.
- Changed `npm run develop` to run `npm run bundle:watch`
- Added `npm run start:patterns`, which will call open a new window and
call `fractal start --sync` and spin up a web server to serve the
pattern library.
- Include `npm run start:patterns` in `npm run develop` in order to
develop on the front end theme assets with the pattern library for
assistance.
- Add `npm run build` to create production ready static assets
(`npm run bundle:prod`) and build a static version of the pattern
library (`npm run build:patterns`) that's moved to `wwwroot\patterns`,
which will make the pattern library accessible when deployed within an
Orchard Core site.

Changed post build events to pre build events. Added pre build event
that will execute `npm run build` while building with "Release"
configuration. This ensures the production ready static assets and
pattern library are included in a release build of the theme.