Add JS to handle enhancing menu component
Added JavaScript to handle a couple of things with regards to the menu.

Firstly when on smaller devices the user can toggle the nav visibility
via a menu button. The `toggleNav` component handles a user clicking
a menu button to toggle the visibility of the nav by adding/removing
a CSS class to the `html` element. Currently there is no CSS to cater
for this.

Lastly the nav can have multiple levels so we've added a `nav`
JavaScript component to show/hide sub nav items.

The JavaScript within the theme was completely empty so we've added a
very basic approach to handle initialising JavaScript components.

(#31)