Add default recipe
A recipe is often required when creating a custom theme in order to
configure the site setup to marry up with what's defined in the theme.
To give a flavour of how a recipe should work and give a handy starting
point the boilerplate now comes with a setup recipe.

Below are some additional changes that contribute to having a default
recipe.

- New parameter for specifying namespace of project when creating a
boilerplate as the recipe needs to add the namespace to the top of the
enabled features collection in the features step.
- Add "Content Menu Item" to allow content editors to build a menu by
selecting content items instead of static urls.
- Override display of `BagPart` to no longer wrap content within a
`section` element.