Use @include rem for font sizes
Searched for and replaced any instance of a font-size: rule and replaced it with our rem mixin, which outputs the rem value and a px backup.
The only elements not changed (html and body) are commented in the CSS already with reasons why they have to remain as percentages/em respectively.