Set theme version to 0.5.0-rc1
We're planning on making a theme created from the boilerplate available
on NuGet so it can be referenced by the site boilerplate. The theme
needs to have a version that represents the version of the theme
boilerplate template package. This will mean any changes to the version
number now needs to happen in a few places.