Refactor CSS component for displaying a menu
CSS linting has been added, which has highlighted an issue with the HTML
& CSS that's used for navigation. There were a lot of
"no-descending-specificity" (see link below) issues within
`list--navigation`that couldn't be removed because of how the CSS was
structured. The primary problem was the use of `list--navigation > li`
for styling the items causing specificity issues that would trigger
the "no-descending-specificity" error.

To resolve these issues and bring the navigation component more in line
with BEM methodology, styles have been moved to a `nav` component. This
makes it's easy to find, easier to extend (adding modifiers) and the CSS
is easier to work with.

The other benefit and the whole reason this change was triggered, the
linting errors are resolved.

https://stylelint.io/user-guide/rules/no-descending-specificity