Support laying out widgets with flow editor
The recent RC2 update of Orchard Core introduced some awesome updates
to the flow editor specifically with positioning widgets with a grid
layout. As there are still some widgets like the `Hero` that need to be
full width, they use breakout styling to ensure they are always the
width of the page.

Currently the `.flow` component is using a really basic grid as
explained in the link below.

https://philipwalton.github.io/solved-by-flexbox/demos/grids/

The conversion of using flows instead of bags for creatable content
types has only been applied to the basic recipe. Plan will be to update
the studio recipe to also use flows. When converting the basic recipe
to use flows, the generic content definitions have been moved to their
own recipe that can be run by all recipes that are included out the box.