Various updates to "Section"
- Add "BackgroundPart" to "Section" widget for defining colour, pattern,
inverted text or whether it's fixed
- Remove "BackgroundImage" field in favour of using "Hero" widget for
displaying a section with content on top of a striking background image
- Add various layout fields for defining padding within section,
spacing between widgets within section flow and vertical alignment.
- Make layout field options consistent and remove any references to
`section` within the underlying values. Doing this gives control to the
HTML & CSS for defining class names instead of having the CSS class
names bleed in to the database.
- Remove `text-and-image` from pattern library
- Remove `section--feature` modifier
- Add CSS variables to `section` for controlling background color/image
and the padding
- Ensure no class modifiers are necessary when "default" is the selected
option on layout fields
- Change `media` example in pattern library to use flows.
- Add additional modifiers to `constrain` and add these additional
options to the content width field on "Section"