Move handling of whether animation repeats to JS
Instead of having modifiers for each animation that dictates whether the
it repeats, the JS will look for data attribute that for controlling
whether the animation repeats. When the data attribute is present and
the value is `true`, the CSS class for trigger animation will be toggled
based on whether the element is in view. By default the animation will
only run once, when the element has CSS class added because it's in
view it'll the JS will remove the obverver meaning the class will never
be removed.