Use CSS variables to drive animations
As we're likely going to want to customise the animations via our theme
settings modules the way the CSS is created needed to be changed to no
longer use a mixin. Added benefit to not using a mixin is the CSS output
is a lot leaner as we're no duplicating styles for each different fade
in animation instead combining them.