Switch from `imagemin-webpack-plugin` to `image-minimizer-webpack-plugin`
Noticed that there was a deprecation warning being outputted when
running the webpack build process, which turned out to be the following
issue.

https://github.com/Klathmon/imagemin-webpack-plugin/issues/116

On the issue, the recommended step to remedy the warning is to migrate
to `image-minimizer-webpack-plugin`, which is an up to date webpack
plugin that is being actively developed.

In the future switching to use Squoosh would be a good move.

https://github.com/webpack-contrib/image-minimizer-webpack-plugin#optimize-with-squoosh