Fix issue when packaging distributable that no CSS file present in new theme
Issue has been introduced since splitting the CSS/JS in to separate
webpack configs. The cause of the issue is the JS config was using a
plugin to delete existing artifacts, which includes the CSS file, which
actually meant it was deleting the CSS artifact created in the current
build. Solution was to make the CSS config responsible for cleaning it's
own artifact folder.