Remove padding from constrain--full
Now that we're outputting constrain--full, it transpires that all constrains have padding by default, which is undesirable if you're putting e.g a full width image into a constrain full. It did make me wonder whether there should be "Full (With Gutter)" and "Full (No Gutter)" as options, so you could do full width stuff but not-quite touch the sides, but in practice I think it's going to always (or almost always) want to be paddingless so let's go with this and see if it comes up.