Switch to Inverted helper class
An adjacent change is being made to the Widgets module that adds an 'inverted' helper class. The modifier classes remain and can still be used for bespoke inversion adaptions if desired, but this change means we can reduce repeated css rules and simplify some of the $class= logic in navs.