Rely on columns helper
Currently this CSS is effectively duplicated in multiple components, and it doesn't need to be (at least, not by default). The helpers/_columns.scss creates the same default behaviour. Content feeds already use it and are doubling up on CSS rules. An adjacent PR is incoming to Widgets to ensure that Gallery can also use the standardised helper classes.