Refactor link and nav
To better support the new system where link styles can be used in the context of a navigation, I've made adjustments here to tackle inadvertent style spilling caused by tag-based CSS and nesting. Adjacent PRs are incoming on Menu and Widgets modules. I also took the opportunity to standardise our approach to link styles, making it more similar to how other components are managed, and reducing the need for repetition between the link component and the default anchor styles.